visuals now work
